add DeviceController config add json tag ,move default function to default.go
